Simple-AXFR test added.
This should test bind and tinydns backends for AXFR bugs.
Uses the test.com zone because it has some nice fancy records.

The downside to this test is that it uses drill (not cleandig) and the order
of the fields is very important. So the gmysql backend will probably need an extra
expected_result.